requirements.txt
asgiref==3.6.0
boto3==1.26.60
botocore==1.29.60
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.4.5
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
environ==1.0
gunicorn==20.0.4
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2020.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.3.1
tzdata==2022.2
urllib3==1.26.13
